feat(delete-event): implement erase All event - add erase all controller
