+++ controllers/events.js
@@ -1,4 +1,6 @@
 
+const { events } = require("../models/index");
+
 var getAllEvents = () => {
 
 };
@@ -12,9 +14,20 @@ var getByActor = () => {
 
 };
 
-
-var eraseEvents = () => {
-
+/**
+ * 
+ * @param {*} req 
+ * @param {*} resp 
+ * @param {*} next 
+ */
+var eraseEvents = async (req, resp, next) => {
+	try {
+		await events.destroy({ truncate: true })
+		resp.statusCode = 200
+		resp.end()
+	} catch (e) {
+		next(e)
+	}
 };
 
 module.exports = {
