+++ controllers/index.js
@@ -1,0 +1,3 @@
+module.exports.actor = require("./actors")
+module.exports.event = require("./events")
+module.exports.repo = require("./repo")
