+++ package.json
@@ -8,7 +8,7 @@
     "pretest": "npm install",
     "test": "./node_modules/.bin/mocha test/index",
     "test:dev": "./node_modules/.bin/mocha test/index",
-    "nodemon": "nodemon app.js"
+    "nodemon": "nodemon ./bin/www"
   },
   "dependencies": {
     "bluebird": "^3.5.0",
