+++ routes/actor.js
@@ -1,6 +1,7 @@
 var express = require('express');
 var router = express.Router();
 
+
 // Routes related to actor.
 
 module.exports = router;
