+++ test/index.js
@@ -1,4 +1,4 @@
- //require("./tests")
-require("./database/actor.spec")
-require("./database/event.spec")
-require("./database/repo.spec")
+require("./tests")
+// require("./database/actor.spec")
+// require("./database/event.spec")
+// require("./database/repo.spec")
